any: restructure calendar
